requirements.txt
aiodns==3.0.0
beautifulsoup4==4.12.2
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cchardet==2.1.7
certifi @ file:///C:/b/abs_55jxithrm3/croot/certifi_1717618077850/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama==0.4.6
dataclasses==0.6
elastic-transport==8.4.0
elasticsearch==8.9.0
et-xmlfile==1.1.0
fake-useragent==1.1.3
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fsspec==2024.6.0
geographiclib==2.0
geopy==2.3.0
gmpy2 @ file:///C:/ci/gmpy2_1645455782955/work
googletransx==2.4.2
huggingface-hub==0.23.4
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib-resources==6.0.0
Jinja2 @ file:///C:/b/abs_92fccttino/croot/jinja2_1716993447201/work
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy==2.0.0
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pillow @ file:///C:/b/abs_95ls_qh4c2/croot/pillow_1714398870615/work
protobuf==5.27.1
PyArabic==0.6.15
pycares==4.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.3.0
pytz==2024.1
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
regex==2024.5.15
requests @ file:///C:/b/abs_9c5n24p1y7/croot/requests_1716902867639/work
safetensors==0.4.3
schedule==1.2.0
scipy==1.13.1
sentencepiece==0.2.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.4.1
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.1
urllib3 @ file:///C:/b/abs_a7hvzm4y95/croot/urllib3_1718912661242/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
XlsxWriter==3.2.0
